More flash fix attempts...
I really hate this thing. It works great on my cheap, but new, BluView4
phone, but I *will* get this stupid thing to work on my Galaxy S5.

The camera on my Galaxy S5 will crash and require a full phone
reboot if an 'enableTorch' call is called while an imageCapture with
flashmode on attempts to take a picture... so now I'm manually turning
on the flash for ImageCapture. This can sometimes mess with image focus
or exposure, as when FlashMode was set on the ImageCapture object the phone
would go through its AutoExposure/AutoFocus routine before snapping a picture.
Turning the flash on for a bit before the capture call seems to help for some
reason but isn't perfect. I don't know why it helps at all because the camera
shouldn't be 'seeing' anything different as long as the flash is on before
the camera, but it seems to help.

I found that when ImageCapture has flash mode on that calling enableTorch
on the CameraControl returned from binding ImageCapture would cause the
ImageCapture to take a snapshot (this is what was crashing my camera:
calling enableTorch(false) would *trigger* ImageCapture to take a picture,
which would try to control the flash and crash). So now when a use case
references the flash it passes along its CameraControl (instead of grabbing
the first non-null CameraControl in the list). With the change to make
ImageCapture use manual flash this shouldn't really have any affect.