Always call enableTorch() when refcnt+
Previously enableTorch(true) would only be called on the first refcnt.
However, sometimes ImageCapture will decide to turn off the flash, which
would require all streams to close to get refcnt down to 0 so that it
could go back up to 1 again to turn back on. By always calling enableTorch,
even when it should already be on, it will allow a new use case to turn
the torch back on (although it must be a new use case, eg a snapshot. Just
opening up new streams without closing existing streams will just share the
same use case and the flash will still be off...)